refactor:  editableFieldModel & readPrettyFieldModel
