+++ packages/core/client/src/flow/models/fields/EditableField/EditableFieldModel.tsx
@@ -16,7 +16,7 @@ import React from 'react';
 import { FormFieldGridModel } from '../..';
 import { ReactiveField } from '../../../formily/ReactiveField';
 import { FieldModel } from '../../base/FieldModel';
-
+import { JsonInput } from '../../common/JsonInput';
 type FieldComponentTuple = [component: React.ElementType, props: Record<string, any>] | any[];
 
 type Structure = {
@@ -24,6 +24,7 @@ type Structure = {
 };
 
 export class EditableFieldModel extends FieldModel<Structure> {
+  static supportedFieldInterfaces = '*' as any;
   field: Field;
   get form() {
     return (this.parent.form as Form) || (this.parent.parent.form as Form);
@@ -34,7 +35,7 @@ export class EditableFieldModel extends FieldModel<Structure> {
   }
 
   get component(): FieldComponentTuple {
-    return [Input, {}];
+    return [JsonInput, {}];
   }
 
   setTitle(title: string) {
