+++ packages/core/client/src/flow/models/fields/ReadPrettyField/ReadPrettyFieldModel.tsx
@@ -13,16 +13,65 @@ import React from 'react';
 import { FieldModel } from '../../base/FieldModel';
 
 export class ReadPrettyFieldModel extends FieldModel {
+  static readonly supportedFieldInterfaces = '*';
+
   getValue() {
     return this.ctx.shared.value;
   }
   @reactive
   public render() {
+    const value = this.getValue();
+    const { prefix, suffix } = this.props;
+    const dataType = this.collectionField?.dataType;
+
+    console.log(dataType, value, this.collectionField.name, this.collectionField.title);
+
+    let content = '';
+
+    if (value === null || value === undefined) {
+      content = ''; // null/undefined 显示为空
+    } else {
+      switch (dataType) {
+        case 'boolean':
+          content = String(value);
+          break;
+        case 'number':
+        case 'double':
+          content = String(value);
+          break;
+        case 'string':
+          content = value;
+          break;
+        case 'object':
+        case 'array':
+          try {
+            content = JSON.stringify(value, null, 2);
+          } catch {
+            content = '[Invalid JSON]';
+          }
+          break;
+        default:
+          // 当 dataType 为空或不识别时根据值类型兜底处理
+          if (typeof value === 'boolean' || typeof value === 'number' || typeof value === 'string') {
+            content = String(value);
+          } else if (typeof value === 'object') {
+            try {
+              content = JSON.stringify(value, null, 2);
+            } catch {
+              content = '[Invalid JSON]';
+            }
+          } else {
+            content = String(value);
+          }
+          break;
+      }
+    }
+
     return (
       <span>
-        {this.props.prefix}
-        {this.translate(this.getValue())}
-        {this.props.suffix}
+        {prefix}
+        {this.translate(content)}
+        {suffix}
       </span>
     );
   }
